requirements.txt
Django==5.0.6
psycopg2-binary==2.9.9
django-allauth==0.63.3
django-taggit==5.0.1
